requirements.txt
pandas>=0.23.0
pyaml==19.4.1
tensorboard==2.0.0
torch==1.4.0
torchvision==0.5.0
